фикс дерева категорий

+++ shipping/models.py
@@ -107,7 +107,7 @@ class ProductUnit(models.Model):
     is_sale = models.BooleanField(default=False)
 
     dimensions = models.JSONField(default=dict)
-    weight = models.IntegerField(default=1)
+    weight = models.FloatField(default=1)
     history_price = models.JSONField(default=list)
     platform_info = models.JSONField(default=dict)
     absolute_profit = models.IntegerField(default=0)
